Reorganized/refactored file uploads and fileOverview
