+++ src/applications/ivc-champva/10-10D/config/form.js
@@ -130,8 +130,8 @@ import { ApplicantAddressCopyPage } from '../../shared/components/applicantLists
 import { hasReq } from '../../shared/components/fileUploads/MissingFileOverview';
 import { fileWithMetadataSchema } from '../../shared/components/fileUploads/attachments';
 
-// import mockData from '../tests/fixtures/data/test-data.json';
-import FileFieldCustom from '../../shared/components/fileUploads/FileUpload';
+import mockData from '../tests/fixtures/data/test-data.json';
+import FileFieldWrapped from '../components/FileUploadWrapper';
 
 /** @type {FormConfig} */
 const formConfig = {
@@ -184,7 +184,7 @@ const formConfig = {
       title: 'Signer information',
       pages: {
         page1: {
-          // initialData: mockData.data,
+          initialData: mockData.data,
           path: 'signer',
           title: 'Which of these best describes you?',
           uiSchema: {
@@ -794,7 +794,7 @@ const formConfig = {
                 ?.relationshipToVeteran === 'child'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantBirthCertUploadUiSchema,
@@ -824,7 +824,7 @@ const formConfig = {
               ) === 'adoption'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantAdoptedUploadUiSchema,
@@ -855,7 +855,7 @@ const formConfig = {
               ) === 'step'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantStepChildUploadUiSchema,
@@ -930,7 +930,7 @@ const formConfig = {
               )
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantSchoolCertUploadUiSchema,
@@ -957,7 +957,7 @@ const formConfig = {
                 'over18HelplessChild'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantHelplessChildUploadUiSchema,
@@ -1077,7 +1077,7 @@ const formConfig = {
                 !get('sponsorIsDeceased', formData))
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantMarriageCertUploadUiSchema,
@@ -1107,7 +1107,7 @@ const formConfig = {
               ) === 'marriedTillDeathRemarriedAfter55'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantSecondMarriageCertUploadUiSchema,
@@ -1141,7 +1141,7 @@ const formConfig = {
               !get('remarriageIsViable', formData?.applicants?.[index])
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantSecondMarriageDivorceCertUploadUiSchema,
@@ -1221,7 +1221,7 @@ const formConfig = {
               ) === 'enrolled'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantMedicarePartAPartBCardsUploadUiSchema,
@@ -1251,7 +1251,7 @@ const formConfig = {
               ) === 'enrolled'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantMedicarePartDCardsUploadUiSchema,
@@ -1281,7 +1281,7 @@ const formConfig = {
               getAgeInYears(formData.applicants[index]?.applicantDOB) >= 65
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: appMedicareOver65IneligibleUploadUiSchema,
@@ -1320,7 +1320,7 @@ const formConfig = {
               get('applicantHasOhi', formData?.applicants?.[index]) === 'yes'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantOhiCardsUploadUiSchema,
@@ -1347,7 +1347,7 @@ const formConfig = {
               ) === 'enrolled'
             );
           },
-          CustomPage: FileFieldCustom,
+          CustomPage: FileFieldWrapped,
           CustomPageReview: null,
           customPageUsesPagePerItemData: true,
           uiSchema: applicantOtherInsuranceCertificationUploadUiSchema,
