+++ src/applications/ivc-champva/10-10D/containers/ConfirmationPage.jsx
@@ -6,10 +6,11 @@ import { connect } from 'react-redux';
 import scrollToTop from 'platform/utilities/ui/scrollToTop';
 import { focusElement } from 'platform/utilities/ui';
 import { VaAlert } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import {
-  requiredFiles,
-  optionalFiles,
-} from '../helpers/supportingDocsVerification';
+// import {
+//   requiredFiles,
+//   optionalFiles,
+// } from '../helpers/supportingDocsVerification';
+import { requiredFiles, optionalFiles } from '../config/requiredUploads';
 import MissingFileOverview from '../../shared/components/fileUploads/MissingFileOverview';
 
 const heading = (
@@ -52,6 +53,7 @@ export function ConfirmationPage(props) {
     showMail: true,
     allPages: form.pages,
     fileNameMap: { ...requiredFiles, ...optionalFiles },
+    requiredFiles,
   });
 
   useEffect(() => {
