+++ src/applications/ivc-champva/shared/utilities.js
@@ -38,3 +38,19 @@ export function makeHumanReadable(inputStr) {
     .map(word => word[0].toUpperCase() + word.substr(1).toLowerCase())
     .join(' ');
 }
+
+/**
+ * Evaluate the `depends` func of each provided page to determine
+ * its value.
+ * @param {object|list} pages A subset of pages within the form
+ * @param {object} data formData used in `depends` calculations
+ * @param {number} index Optional argument to pass to `depends` if evaluating list and loop page `depends`
+ * @returns A filtered list of pages where `depends` was true
+ */
+export function getConditionalPages(pages, data, index) {
+  const tmpPg =
+    typeof pages === 'object' ? Object.keys(pages).map(pg => pages[pg]) : pages;
+  return tmpPg.filter(
+    pg => pg.depends === undefined || pg?.depends({ ...data }, index),
+  );
+}
